scrap-code/bot-weather/requirements.txt
pyowm==3.3.0
pyTelegramBotAPI==4.8.0
python-dotenv==0.21.0
